pyproject.toml
[project]
name = "agentsflow"
version = "0.1.1"
description = "Easily create & connect autonomous autogen AI agents from a streamlined interface."
authors = [
    {name = "jaemil", email = "emilhopf@gmail.com"},
    {name = "TomExMachina", email = "tom@firesight.ai"},
]
dependencies = [
    "fastapi>=0.103.2",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.23.2",
    "jinja2>=3.1.2",
    "pyautogen[blendsearch]>=0.1.7",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "GPL-3.0-only"}

